requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
pandas
python-dotenv